katana installs only!
added a loop that checks in the directory that checks for the name 'Katana'. if Katana is present in the file name, it's added to the list of installs in the dropdown menu. This is to prevent things like modo/mari/nuke/anything else that's in that folder being listed in the dropdown.